+++ changelog.md
@@ -11,7 +11,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Added
 
 - Add CDP integration tests to CI (`2da6a4f`)
-- Implement a `CollisionPolicy` rather than a hard-coded gap in `make_draggable()` (``)
+- Implement a `CollisionPolicy` rather than a hard-coded gap in `make_draggable()` (`b472e95`)
+
+## Fixed
+
+- Adjust overlap resolver to make non-overlap a best guess, not a guarantee (``)
 
 # [Released]